requirements.txt
streamlit
python-dotenv
faiss-cpu
pypdf
langchain
langchain_text_splitters
langchain-community
langchain-google-genai
google-generativeai
sentence-transformers
pandas
numpy
rapidfuzz
faster-whisper
soundfile